bot/requirements.txt
python-dotenv
aiohttp
PyNaCl
ffmpeg-python
yt-dlp
youtube-search-python
wget
async-timeout
pytz
openai
humanize
requests
Pillow
emoji
aiosqlite
scipy
numpy
wave
pyfiglet
setuptools
anthropic
discord.py==2.3.2
uuid
googletrans
textblob
colorama
